# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lakeview",
    "version": "0.3.0"
  },
  "paths": {
    "/api/roots": {
      "get": {
        "summary": "Get Roots",
        "operationId": "get_roots_api_roots_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/file/{root}/{path}": {
      "get": {
        "summary": "Get File",
        "description": "Serve raw file bytes from any root, for preview in the frontend.",
        "operationId": "get_file_api_file__root___path__get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets": {
      "get": {
        "summary": "Get Datasets",
        "operationId": "get_datasets_api_datasets_get",
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/d/{root}/{path}/info": {
      "get": {
        "summary": "Get Info",
        "operationId": "get_info_api_d__root___path__info_get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/d/{root}/{path}/schema": {
      "get": {
        "summary": "Get Schema",
        "operationId": "get_schema_api_d__root___path__schema_get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/d/{root}/{path}/rows": {
      "get": {
        "summary": "Get Rows",
        "operationId": "get_rows_api_d__root___path__rows_get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericRowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/d/{root}/{path}/row/{offset}": {
      "get": {
        "summary": "Get Row",
        "operationId": "get_row_api_d__root___path__row__offset__get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Row Api D  Root   Path  Row  Offset  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/d/{root}/{path}/blob/{offset}/{column}": {
      "get": {
        "summary": "Get Blob",
        "description": "Stream one cell's raw bytes with a detected Content-Type.\n\nCovers v1 / v2 Lance blob columns and plain binary columns. Content-Type\nis taken from the blob's source URI suffix when available (v2 URI-refs),\nand falls back to magic-byte sniffing of the first bytes. Non-binary\ncolumns 404.",
        "operationId": "get_blob_api_d__root___path__blob__offset___column__get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offset"
            }
          },
          {
            "name": "column",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Column"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/d/{root}/{path}/view": {
      "get": {
        "summary": "Get View",
        "operationId": "get_view_api_d__root___path__view_get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/d/{root}/{path}/view/{key}": {
      "get": {
        "summary": "Get View Detail",
        "operationId": "get_view_detail_api_d__root___path__view__key__get",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentRunDetail": {
        "properties": {
          "row": {
            "additionalProperties": true,
            "type": "object",
            "title": "Row"
          },
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "row",
          "messages"
        ],
        "title": "AgentRunDetail"
      },
      "AgentRunSidebar": {
        "properties": {
          "row_offset": {
            "type": "integer",
            "title": "Row Offset"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "output": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "metadata": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "row_offset"
        ],
        "title": "AgentRunSidebar"
      },
      "AgentRunStats": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "ok": {
            "type": "integer",
            "title": "Ok"
          },
          "wrong": {
            "type": "integer",
            "title": "Wrong"
          },
          "error": {
            "type": "integer",
            "title": "Error"
          },
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          }
        },
        "type": "object",
        "required": [
          "total",
          "ok",
          "wrong",
          "error",
          "pending"
        ],
        "title": "AgentRunStats"
      },
      "ColumnInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable"
          },
          "is_blob": {
            "type": "boolean",
            "title": "Is Blob",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "nullable"
        ],
        "title": "ColumnInfo"
      },
      "DatasetEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "kind"
        ],
        "title": "DatasetEntry"
      },
      "DatasetInfoResponse": {
        "properties": {
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnInfo"
            },
            "type": "array",
            "title": "Columns"
          },
          "plugin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plugin"
          },
          "filters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filters",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "row_count",
          "columns"
        ],
        "title": "DatasetInfoResponse"
      },
      "DatasetListResponse": {
        "properties": {
          "root": {
            "type": "string",
            "title": "Root"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetEntry"
            },
            "type": "array",
            "title": "Datasets"
          }
        },
        "type": "object",
        "required": [
          "root",
          "path",
          "datasets"
        ],
        "title": "DatasetListResponse"
      },
      "GenericRowListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "rows"
        ],
        "title": "GenericRowListResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PluginDetailResponse": {
        "properties": {
          "plugin": {
            "type": "string",
            "title": "Plugin"
          },
          "data": {
            "$ref": "#/components/schemas/AgentRunDetail"
          }
        },
        "type": "object",
        "required": [
          "plugin",
          "data"
        ],
        "title": "PluginDetailResponse"
      },
      "PluginViewResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/AgentRunSidebar"
            },
            "type": "array",
            "title": "Rows"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRunStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "plugin": {
            "type": "string",
            "title": "Plugin"
          }
        },
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "rows",
          "plugin"
        ],
        "title": "PluginViewResponse"
      },
      "RootInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "uri": {
            "type": "string",
            "title": "Uri"
          }
        },
        "type": "object",
        "required": [
          "name",
          "uri"
        ],
        "title": "RootInfo"
      },
      "RootsResponse": {
        "properties": {
          "roots": {
            "items": {
              "$ref": "#/components/schemas/RootInfo"
            },
            "type": "array",
            "title": "Roots"
          },
          "default": {
            "type": "string",
            "title": "Default"
          }
        },
        "type": "object",
        "required": [
          "roots",
          "default"
        ],
        "title": "RootsResponse"
      },
      "SchemaResponse": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnInfo"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "columns"
        ],
        "title": "SchemaResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}